# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pastebin.fi API",
    "description": "Some endpoints might return more fields than are specified in this document. They are not supposed to be used by public until they are added to this document (since they are mostly for testing and can be removed at any point of time).\n\nUseful resources:\n- [Pastebin.fi API backend code](https://github.com/pastebin-fi/PowerPaste)\n- [Official frontend for pastebin.fi API](https://github.com/pastebin-fi/frontend)",
    "termsOfService": "https://pastebin.fi/about",
    "contact": {
      "email": "sysadmin@pastebin.fi"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.pastebin.fi",
      "description": "Production server (uses live data)"
    }
  ],
  "tags": [
    {
      "name": "paste",
      "description": "Requests related to pastes"
    },
    {
      "name": "metrics",
      "description": "Requestes related to metrics"
    }
  ],
  "paths": {
    "/metrics": {
      "get": {
        "tags": [
          "metrics"
        ],
        "summary": "Get metrics",
        "description": "Get public metrics of the service in JSON format",
        "operationId": "getMetrics",
        "responses": {
          "200": {
            "description": "JSON object containing metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetricsResponse"
                }
              }
            }
          },
          "404": {
            "description": "No metrics available (no pastes created / database query failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "title": "Liitteitä ei löytynyt",
                  "message": "Emme ole vastaanottaneet vielä yhtäkään liitettä. Ole ensimmäinen!"
                }
              }
            }
          }
        }
      }
    },
    "/pastes": {
      "get": {
        "tags": [
          "paste"
        ],
        "summary": "Get list of public pastes",
        "description": "Get list of public pastes",
        "operationId": "listPastes",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search for the following text in title and paste (keep in mind that the title is weighted)"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "The number of pastes to skip before starting to collect the result set"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 0,
              "maximum": 30
            },
            "description": "The amount of pastes that are returned in the result set."
          },
          {
            "in": "query",
            "name": "sorting",
            "schema": {
              "type": "string",
              "default": "-date",
              "enum": [
                "date",
                "-date",
                "meta.views",
                "-meta.views",
                "meta.size",
                "-meta.size"
              ]
            },
            "description": "Sort the results from mongodb with the following method. Hyphen (-) in front of the value switch the sorting to descending order (normally ascending). Has to match the available values or else set to default value."
          }
        ],
        "responses": {
          "200": {
            "description": "Paste list found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasteListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "paste"
        ],
        "summary": "Create a paste",
        "description": "This endpoint is ratelimited to 20 pastes per 30 mins (this might change in the future when accounts come).",
        "requestBody": {
          "description": "Optional description in *Markdown*",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paste has been created succesfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasteCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request body missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "title": "Pyynnön vartalo on pakollinen",
                  "message": "Et voi luoda tyhjää pyyntöä."
                }
              }
            }
          },
          "403": {
            "description": "Blocked by safeguards\n\nCurrent safeguards:\n  - AbuseIPDB check\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "title": "Pääsy hylätty",
                  "message": "IP-osoitteesi maine on huono, joten hylkäsimme pyyntösi uuden liitteen luomiseksi."
                }
              }
            }
          },
          "409": {
            "description": "The same content is in another paste",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "title": "Liite on jo olemassa",
                  "description": "Luotu liite on jo olemassa, joten sitä ei luotu.",
                  "data": {
                    "pasteIdentifier": "Daehn4t"
                  }
                }
              }
            }
          },
          "413": {
            "description": "Exceeds limits defined by server\n\nLimits:\n  - Programming language must be under 30 characters\n  - Paste size must be under 10 MB (Megabytes)\n  - Title must be under 300 characters\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "large": {
                    "summary": "Over 10MB pastes are not allowed",
                    "value": {
                      "title": "Liite on liian iso",
                      "message": "Palveluun ei voi luoda yli kymmenen (10) MB liitteitä."
                    }
                  },
                  "name": {
                    "summary": "Names over 300 characters are disallowed",
                    "value": {
                      "title": "Virheellinen nimi",
                      "message": "Palveluun ei voi luoda liitettä yli 300 merkin otsikolla."
                    }
                  },
                  "language": {
                    "summary": "Programming language names over 30 characters are disallowed",
                    "value": {
                      "title": "Virheellinen ohjelmointikieli",
                      "message": "Ohjelmointikielen nimi ei voi olla yli kolmeakymmentä (30) merkkiä."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unknown error occurred when creating the paste (database connection might have broken)\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "title": "Liitettä ei voitu luoda",
                  "description": "Jotain meni vikaan, eikä liitettä luotu."
                }
              }
            }
          }
        }
      }
    },
    "/pastes/{id}": {
      "get": {
        "tags": [
          "paste"
        ],
        "summary": "Get paste",
        "description": "Get paste and its metadata (views & size)",
        "operationId": "getPaste",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The paste ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Paste found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Paste not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "title": "Liitettä ei löytynyt",
                  "message": "Olemme pahoillamme, mutta hakemaasi liitettä ei löytynyt"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PasteListResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string",
              "description": "Paste title"
            },
            "id": {
              "type": "string",
              "description": "Paste ID (very useful for querying it later)"
            },
            "meta": {
              "type": "object",
              "properties": {
                "views": {
                  "type": "integer",
                  "description": "Amount of page loads the paste has received"
                },
                "size": {
                  "type": "integer",
                  "format": "byte",
                  "description": "Size in bytes"
                }
              }
            },
            "date": {
              "type": "string",
              "format": "datetime",
              "description": "Time when paste was created"
            }
          },
          "example": [
            {
              "title": "jamsanseutu.fi - 2000009558326 - Maksumuuri poistettu",
              "id": "GOtkjL0",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 2,
                "size": 1929
              },
              "date": "2023-05-05T05:58:52.590Z"
            },
            {
              "title": "jamsanseutu.fi - 2000009560587 - Maksumuuri poistettu",
              "id": "MOaY1Ux",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 4,
                "size": 7064
              },
              "date": "2023-05-05T05:49:57.027Z"
            },
            {
              "title": "Ecoshift Corp, LED Tube Lights Manila Philippines",
              "id": "fgJRrSP",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 7,
                "size": 359
              },
              "date": "2023-05-04T11:00:18.642Z"
            },
            {
              "title": "hs.fi - 2000009511830 - Maksumuuri poistettu",
              "id": "WQ9LqBe",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 16,
                "size": 16650
              },
              "date": "2023-05-03T07:32:11.872Z"
            },
            {
              "title": "LinkHelpers Phoenix Digital Marketing",
              "id": "1uw3d0o",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 12,
                "size": 506
              },
              "date": "2023-05-03T06:15:41.181Z"
            },
            {
              "title": "aamulehti.fi - 2000008948411 - Maksumuuri poistettu",
              "id": "RDmebqX",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 12,
                "size": 3382
              },
              "date": "2023-05-02T12:51:29.109Z"
            },
            {
              "title": "AI Monthly SEO plans",
              "id": "IZs1kdC",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 11,
                "size": 398
              },
              "date": "2023-05-02T05:31:15.515Z"
            },
            {
              "title": "Improve local visibility for pastebin.fi",
              "id": "3iNenOB",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 11,
                "size": 440
              },
              "date": "2023-04-30T12:52:16.058Z"
            },
            {
              "title": "hs.fi - 2000009514770 - Maksumuuri poistettu",
              "id": "rVTg2Qv",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 23,
                "size": 37275
              },
              "date": "2023-04-30T11:46:01.626Z"
            },
            {
              "title": "",
              "id": "HRvvPWw",
              "author": "Tuntematon lataaja",
              "meta": {
                "votes": null,
                "favs": null,
                "views": 23,
                "size": 1209
              },
              "date": "2023-04-29T12:24:03.176Z"
            }
          ]
        }
      },
      "PasteCreateRequest": {
        "type": "object",
        "required": [
          "title",
          "paste"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "Placeholder text",
            "description": "Paste title"
          },
          "paste": {
            "type": "string",
            "example": "Lorem ipsum dolor sit amet...",
            "description": "Paste content"
          },
          "language": {
            "type": "string",
            "default": "text",
            "example": "python",
            "description": "Paste programming langauge (currently not used anywhere)"
          },
          "private": {
            "type": "boolean",
            "default": false,
            "example": true,
            "description": "Do not show the paste to others (only the ones that have access to the ID can see it)"
          }
        }
      },
      "PasteCreateResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "Placeholder text",
            "description": "Paste identifier can be used for a redirect to the paste when created"
          }
        }
      },
      "PasteMeta": {
        "required": [
          "views",
          "size"
        ],
        "type": "object",
        "properties": {
          "views": {
            "type": "integer",
            "example": 1,
            "description": "Amount of page loads the paste has received"
          },
          "size": {
            "type": "integer",
            "format": "byte",
            "example": 109,
            "description": "Size in bytes"
          }
        }
      },
      "PasteResponse": {
        "required": [
          "meta",
          "title",
          "id",
          "hidden",
          "date"
        ],
        "type": "object",
        "properties": {
          "meta": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PasteMeta"
            }
          },
          "title": {
            "type": "string",
            "example": "Example title"
          },
          "id": {
            "type": "string",
            "example": "58hMQlk"
          },
          "hidden": {
            "type": "boolean",
            "example": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-03T11:19:20.137Z"
          }
        }
      },
      "ErrorResponse": {
        "required": [
          "title",
          "message"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "error type",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "example": "",
            "description": "More specific data of the error"
          },
          "data": {
            "type": "object",
            "description": "In some cases additional fields are given in the error response",
            "properties": {
              "pasteIdentifier": {
                "type": "string",
                "description": "existing paste ID"
              }
            }
          }
        }
      },
      "MetricsResponse": {
        "required": [
          "totalViews",
          "pasteCount - total - public - private"
        ],
        "type": "object",
        "properties": {
          "totalviews": {
            "type": "integer",
            "example": 138953,
            "description": "Sum of all paste views"
          },
          "pasteCount": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "description": "Amount of all pastes (both public and private)",
                "example": 1557
              },
              "public": {
                "type": "integer",
                "description": "Amount of public pastes",
                "example": 1129
              },
              "private": {
                "type": "integer",
                "description": "Amount of private pastes",
                "example": 428
              }
            }
          }
        }
      }
    }
  }
}